Create PowerOfTwo.java
Intuition :
To determine if a number 𝑛 is a power of two, recognize that powers of two have a specific property: they can be repeatedly divided by 2 until they become 1. If at any step during this process the number becomes odd and is not 1, then the number is not a power of two.

Approach :
This is a beginner friendly approach so that its very easy to understand.

1.Check if n is 1, which is 2^0.
2.If n is odd and greater than 1, it cannot be a power of two.
3.Loop to repeatedly divide n by 2.
4.If n becomes 1, it is a power of two.
5.If n becomes odd (and not 1), it cannot be a power of two.
6.If n never becomes 1, it is not a power of two.

Time complexity:
The time complexity of the code is O(log n).

Space complexity:
The space complexity of the code is O(1).